requirements.txt
django
django-rest-framework
xmltodict==0.13.0
black
